refactor: Reorganize blog following cesium utils pattern
