+++ apps/web/app/blog/_components/table-of-contents.tsx
@@ -11,7 +11,7 @@ import { cn } from "@juun/ui/lib/utils";
 import { ArrowRight, Bookmark } from "lucide-react";
 import { Fragment, useEffect, useState } from "react";
 
-import { getHeadings, Heading } from "@/app/blog/utils";
+import { getHeadings, Heading } from "../_utils/post";
 
 interface TableOfContentsProps {
   /**
