+++ apps/web/app/blog/_data/bundle-optimization.tsx
@@ -17,7 +17,7 @@ import Image from "next/image";
 import Link from "next/link";
 import { Fragment, Suspense } from "react";
 
-import { BlogHeader, BlogHeaderSkeleton } from "@/components/blog/header";
+import { BlogHeader, BlogHeaderSkeleton } from "@/app/blog/_components/header";
 
 export const metadata = {
   title: "Bundle Optimization",
