--- apps/web/app/blog/types.ts
@@ -1,3 +1,0 @@
-import { BaseMetadata } from "@/types/post.types";
-
-export type BlogMetadata = BaseMetadata & { wordCount?: number };
